quiz-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
pika==1.3.2